update pb parser to avoid drainer failure
